ATR-838: renamed test file, folder and namespace
