+++ src/Acuminator/Acuminator.Tests/Acuminator.Tests.csproj
@@ -90,11 +90,11 @@
     <EmbeddedResource Include="Tests\StaticAnalysis\BqlParameterMismatch\Sources\FbqlCalls.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\BqlParameterMismatch\Sources\BqlWithGenericDacFieldCall.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\ForbiddenFieldsInDac\Sources\DacFieldsWithCompanyPrefix.cs" />
-    <Compile Include="Tests\StaticAnalysis\AsyncVoidMethods\AsyncVoidMethodsTests.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethods\Sources\Methods\PartialVoidAsyncMethods.OtherDeclaration.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethods\Sources\Methods\PartialVoidAsyncMethods.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethods\Sources\Methods\RegularVoidAsyncMethods.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethods\Sources\Methods\NormalAsyncMethods.cs" />
+    <Compile Include="Tests\StaticAnalysis\AsyncVoidMethodsAndLambdas\AsyncVoidMethodsAndLambdasTests.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethodsAndLambdas\Sources\Methods\PartialVoidAsyncMethods.OtherDeclaration.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethodsAndLambdas\Sources\Methods\PartialVoidAsyncMethods.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethodsAndLambdas\Sources\Methods\RegularVoidAsyncMethods.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\AsyncVoidMethodsAndLambdas\Sources\Methods\NormalAsyncMethods.cs" />
     <Compile Include="Tests\StaticAnalysis\BannedApi\BannedApiTestsIsv.cs" />
     <Compile Include="Tests\StaticAnalysis\BannedApi\BannedApiAnalysisDisabledTests.cs" />
     <Compile Include="Tests\StaticAnalysis\BannedApi\BannedApiTestsIsvWithExternalFiles.cs" />
